services/api-gateway/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
kafka-python==2.0.2
redis==5.0.1
slowapi==0.1.9
prometheus-client==0.20.0
pydantic==2.6.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
httpx==0.27.0
pytest==8.1.0
pytest-asyncio==0.23.5
pytest-mock==3.12.0
